AMPET-148-AuditLoggerCleanerJob

+++ amp/WEB-INF/src/org/digijava/module/aim/services/auditcleaner/AuditCleanerJob.java
@@ -24,9 +24,11 @@ public class AuditCleanerJob extends ConnectionCleaningJob {
     @Override
     public void executeInternal(JobExecutionContext contex) throws JobExecutionException {
         String deletetime = FeaturesUtil.getGlobalSettingValue(GlobalSettingsConstants.AUTOMATIC_AUDIT_LOGGER_CLEANUP);
+        if (!(Integer.parseInt(deletetime) == -1)) {
         logger.info("Starting to delete audit logs older than " + deletetime + " days.............................");
         AuditLoggerUtil.deleteLogsByPeriod(deletetime);
         logger.info("Delete Audit logs finished...................................................");
+        }
     }
 
 }
